cast return

+++ demos/projects/ST/b-l475e-iot01a/sample_gsg_device.c
@@ -196,6 +196,6 @@ uint32_t ulCreateTelemetry( uint8_t * pucTelemetryData,
 
     LogInfo( ( "Telemetry message sent %.*s\r\n", lBytesWritten, pucTelemetryData ) );
 
-    return lBytesWritten;
+    return (uint32_t) lBytesWritten;
 }
 /*-----------------------------------------------------------*/
